refactor: separate shared construction and read ownership
