+++ cmd/internal/serverapp/graph_composition.go
@@ -1,0 +1,7 @@
+package serverapp
+
+import "github.com/markhuangai/dense-mem/internal/service/graphview"
+
+func buildGraphApplication(store graphview.SemanticStore) graphview.Service {
+	return graphview.NewSemantic(store)
+}
