+++ cmd/internal/serverapp/trace_composition.go
@@ -1,0 +1,7 @@
+package serverapp
+
+import "github.com/markhuangai/dense-mem/internal/service/contextservice"
+
+func buildContextApplication(store contextservice.SemanticTraceStore) contextservice.Service {
+	return contextservice.NewSemantic(store)
+}
